pyproject.toml
[tool.poetry]
name = "cta-stop-watch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cta-stop-watch"}]

[tool.poetry.dependencies]
python = "^3.10"
datetime = "^5.5"
polars = "^1.1.0"
pathlib = "^1.0.1"
duckdb = "^1.0.0"
pandas = "^2.2.2"
geopandas = "^1.0.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyarrow = "^16.1.0"
folium = "^0.16.0"
black = "^24.4.2"
matplotlib = "^3.9.0"
ruff = "^0.4.8"
mapclassify = "^2.6.1"
env = "^0.1.0"
seaborn = "^0.13.2"
gtfs-kit = "^6.1.0"
argparse = "^1.4.0"
branca = "^0.7.2"
plotly = "^5.22.0"
nbformat = "^5.10.4"
selenium = "^4.22.0"
geckodriver = "^0.0.1"
pyproj = "^3.6.1"
contextily = "^1.6.0"
shapely = "^2.0.5"
fastapi = "^0.111.1"
jinja2 = "^3.1.4"
uvicorn = "^0.30.3"
google-cloud-storage = "^2.18.2"
pendulum = "^3.0.0"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
